add recipedetail module

+++ app/src/main/java/com/xiongxh/baking_app/recipes/RecipesContract.java
@@ -18,7 +18,7 @@ public interface RecipesContract {
     interface Presenter extends BasePresenter<View> {
         //void loadRecipesFromRepository(boolean forceUpdate, RecipesIdlingResource resource);
         void loadRecipes();
-        void openRecipeDetails(long recipeId);
+        void openRecipeDetails(int recipeId);
 
         //void syncData();
